Updated namespaces from folder renaming.
