+++ Tricycle.UI/Pages/ConfigPage.xaml
@@ -2,8 +2,8 @@
 <ContentPage
     xmlns="http://xamarin.com/schemas/2014/forms"
     xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
-    xmlns:local="clr-namespace:Tricycle.UI.Controls"
-    x:Class="Tricycle.UI.Views.ConfigPage"
+    xmlns:local="clr-namespace:Tricycle.UI.Views"
+    x:Class="Tricycle.UI.Pages.ConfigPage"
     BackgroundColor="{StaticResource BackgroundColor}">
     <ContentPage.Resources>
         <OnPlatform x:Key="ButtonRowHeight" x:TypeArguments="GridLength">
