+++ Tricycle.UI/Pages/MainPage.xaml
@@ -1,8 +1,8 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <ContentPage xmlns="http://xamarin.com/schemas/2014/forms"
              xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
-             xmlns:local="clr-namespace:Tricycle.UI.Controls"
-             x:Class="Tricycle.UI.Views.MainPage"
+             xmlns:local="clr-namespace:Tricycle.UI.Views"
+             x:Class="Tricycle.UI.Pages.MainPage"
              BackgroundColor="{StaticResource BackgroundColor}">
     <ContentPage.Resources>
         <OnPlatform x:Key="PanelWidth" x:TypeArguments="FlexBasis">
