+++ Tricycle.UI/Pages/MainPage.xaml.cs
@@ -11,7 +11,7 @@ using Tricycle.UI.ViewModels;
 using Tricycle.Utilities;
 using Xamarin.Forms;
 
-namespace Tricycle.UI.Views
+namespace Tricycle.UI.Pages
 {
     // Learn more about making custom code visible in the Xamarin.Forms previewer
     // by visiting https://aka.ms/xamarinforms-previewer
