+++ Tricycle.UI/Pages/PreviewPage.xaml
@@ -1,8 +1,8 @@
 ﻿<?xml version="1.0" encoding="UTF-8"?>
 <ContentPage xmlns="http://xamarin.com/schemas/2014/forms"
              xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
-             xmlns:local="clr-namespace:Tricycle.UI.Controls"
-             x:Class="Tricycle.UI.Views.PreviewPage"
+             xmlns:local="clr-namespace:Tricycle.UI.Views"
+             x:Class="Tricycle.UI.Pages.PreviewPage"
              BackgroundColor="{StaticResource BackgroundColor}">
     <StackLayout Margin="10, 10, 10, 15">
         <FlexLayout JustifyContent="End">
