+++ Tricycle.UI/Views/AudioConfigSection.xaml
@@ -1,8 +1,8 @@
 ﻿<?xml version="1.0" encoding="UTF-8"?>
 <ContentView xmlns="http://xamarin.com/schemas/2014/forms"
              xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
-             xmlns:local="clr-namespace:Tricycle.UI.Controls"
-             x:Class="Tricycle.UI.Controls.AudioConfigSection">
+             xmlns:local="clr-namespace:Tricycle.UI.Views"
+             x:Class="Tricycle.UI.Views.AudioConfigSection">
     <ContentView.Resources>
         <x:Double x:Key="QualityWidth">75</x:Double>
     </ContentView.Resources>
