+++ Tricycle.UI/Views/FramedPanel.xaml
@@ -1,7 +1,7 @@
 ﻿<?xml version="1.0" encoding="UTF-8"?>
 <ContentView xmlns="http://xamarin.com/schemas/2014/forms"
              xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
-             x:Class="Tricycle.UI.Controls.FramedPanel">
+             x:Class="Tricycle.UI.Views.FramedPanel">
     <ContentView.ControlTemplate>
         <ControlTemplate>
             <Frame BackgroundColor="{StaticResource PanelBackgroundColor}"
