+++ Tricycle.UI/Views/FramedPanel.xaml.cs
@@ -1,7 +1,7 @@
 ﻿using System.Runtime.CompilerServices;
 using Xamarin.Forms;
 
-namespace Tricycle.UI.Controls
+namespace Tricycle.UI.Views
 {
     public partial class FramedPanel : ContentView
     {
