+++ Tricycle.UI/Views/Pill.xaml
@@ -1,7 +1,7 @@
 ﻿<?xml version="1.0" encoding="UTF-8"?>
 <ContentView xmlns="http://xamarin.com/schemas/2014/forms"
              xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
-             x:Class="Tricycle.UI.Controls.Pill"
+             x:Class="Tricycle.UI.Views.Pill"
              BackgroundColor="#3366ff"
              Padding="1">
     <ContentView.Content>
