+++ Tricycle.UI/Views/Pill.xaml.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.Runtime.CompilerServices;
 using Xamarin.Forms;
 
-namespace Tricycle.UI.Controls
+namespace Tricycle.UI.Views
 {
     public partial class Pill : ContentView
     {
