+++ Tricycle.UI/Views/PlatformButton.xaml
@@ -3,9 +3,9 @@
              xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
              xmlns:d="http://xamarin.com/schemas/2014/forms/design"
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-             xmlns:controls="clr-namespace:Tricycle.UI.Controls"
+             xmlns:controls="clr-namespace:Tricycle.UI.Views"
              mc:Ignorable="d"
-             x:Class="Tricycle.UI.Controls.PlatformButton">
+             x:Class="Tricycle.UI.Views.PlatformButton">
   <ContentView.Content>
       <StackLayout>
             <OnPlatform x:TypeArguments="View">
