+++ Tricycle.UI/Views/SnappableSlider.cs
@@ -2,7 +2,7 @@
 using System.Runtime.CompilerServices;
 using Xamarin.Forms;
 
-namespace Tricycle.UI.Controls
+namespace Tricycle.UI.Views
 {
     public class SnappableSlider : Slider
     {
